+++ src/main/java/com/brunoandreotti/gametracker/dtos/gametrack/GameTrackResponseDTO.java
@@ -15,7 +15,7 @@ public class GameTrackResponseDTO {
 
     GameStatus status ;
 
-    Integer rate;
+    Double rate;
 
     Double hoursPlayed;
